+++ src/rest/services/index.ts
@@ -1,1 +1,3 @@
 export * from './mockLogger/mockLogger.interface';
+export * from './advancedCache/advancedCache.service';
+export * from './advancedCache/advancedCache.interface';